Move python files under honeybadgermpc pkg dir
